test(e2e): avoid write to committed file (#3046)
